+++ LibSrc/GlkStream.m
@@ -50,7 +50,7 @@
 }
 
 - (void) dealloc {
-	NSLog(@"GlkStream dealloc %x", self);
+	NSLog(@"GlkStream dealloc %x", (unsigned int)self);
 	
 	if (inlibrary)
 		[NSException raise:@"GlkException" format:@"GlkStream reached dealloc while in library"];
